requirements.txt
nltk
fuzzy
pyspellchecker
tk
flask
flask-cors
